Don't send feature event unless we should (#8)
Previously the Rust SDK was always sending a full "kind": "feature" event for every evaluation, in addition to the summary events. This disables those events, unless either the flag-level trackEvents property was set, or the evaluation hit a rule or fallthrough that had their respective trackEvents properties set.

I also implemented the very-special logic whereby we sometimes send reasons even if the caller didn't ask for them, and refactored the summary event logic to make it clearer that it only summarises feature events.