Improve event processor spec alignment (#45)
The add and flush methods are not supposed to block or return success or
failure condition. However, if an event exceeds the capacity of the
event queue, we are supposed to notify the user this has occurred one
time.

Originally we were triggering flushes based off a batch size and also
some max batch deadline. The spec requires neither of these, so I have
removed them. This addresses the issue of requiring additional
configuration and takes us one step closer to alignment in this
processor.

I also did away with the separate flush channel. The spec requires only
a single inbox channel which includes messages for both events and flush
requests.