Remove locking when possible (#52)
We can remove locking from the event processor and the data source as it
doesn't mutate it's own state, and it is fully owned by the client.

The data store both mutates it's own internal state and is shared
between the data source and the client. We cannot remove the locking
functionality, but we can replace the Mutex with a RwLock
implementation.

We opt for using the parking_lot::RwLock instead of the
std::sync::RwLock because the built-in one on Linux suffers from writer
starvation whereas the parking_lot implementation offers a fair locking
mechanism.