fix: Limit triomphe dependency to <=0.1.10
Our SDK uses `moka` for some unit tests, which in turn has a dependency on
`triomphe`. Neither of these crates provide an MSRV, resulting in their
latest release breaking on rustc 1.74.

To avoid bumping our own MSRV to 1.76 for now, we are going to force the
triomphe dependency to be <=0.1.10. Once we bump rustc in the future, we
can clean up this change.